Move interlocked tests to C#9
